fix: separate entiites by domain
