+++ src/main/java/com/example/study_monster_back/board/entity/Board.java
@@ -1,9 +1,11 @@
-package com.example.study_monster_back.entity;
-
+package com.example.study_monster_back.board.entity;
 
 import java.time.LocalDateTime;
 
 import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import com.example.study_monster_back.user.entity.User;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -15,15 +17,19 @@ import lombok.Data;
 
 @Entity
 @Data
-public class StudyMember{
+public class Board{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    private String title;
+    private String content;
     @CreatedDate
-    private LocalDateTime joined_at;
+    private LocalDateTime created_at;
+    @LastModifiedDate
+    private LocalDateTime updated_at;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
-    @ManyToOne(fetch = FetchType.LAZY)
-    private StudyGroup studyGroup;
+    
 }
