+++ src/main/java/com/example/study_monster_back/comment/entity/Comment.java
@@ -1,9 +1,12 @@
-package com.example.study_monster_back.entity;
+package com.example.study_monster_back.comment.entity;
 
 import java.time.LocalDateTime;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
+
+import com.example.study_monster_back.board.entity.Board;
+import com.example.study_monster_back.user.entity.User;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -15,19 +18,19 @@ import lombok.Data;
 
 @Entity
 @Data
-public class Board{
+public class Comment{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String title;
     private String content;
     @CreatedDate
     private LocalDateTime created_at;
     @LastModifiedDate
     private LocalDateTime updated_at;
-
+    
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
-    
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Board board;
 }
