+++ src/main/java/com/example/study_monster_back/group/entity/StudyGroup.java
@@ -1,11 +1,13 @@
-package com.example.study_monster_back.entity;
+package com.example.study_monster_back.group.entity;
+
 
 import java.time.LocalDateTime;
 
 import org.springframework.data.annotation.CreatedDate;
 
+import com.example.study_monster_back.user.entity.User;
+
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -14,16 +16,19 @@ import lombok.Data;
 
 @Entity
 @Data
-public class Feedback{
+public class StudyGroup{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String content;
+    private String name;
     @CreatedDate
     private LocalDateTime created_at;
+    private String description;
+    private Integer limit_members;
+    private LocalDateTime deadline;
+    private String status;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Board board;
-    
+    @ManyToOne
+    private User creator;
 }
