+++ src/main/java/com/example/study_monster_back/group/entity/StudyMember.java
@@ -1,11 +1,14 @@
-package com.example.study_monster_back.entity;
+package com.example.study_monster_back.group.entity;
 
 
 import java.time.LocalDateTime;
 
 import org.springframework.data.annotation.CreatedDate;
 
+import com.example.study_monster_back.user.entity.User;
+
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -14,19 +17,15 @@ import lombok.Data;
 
 @Entity
 @Data
-public class StudyGroup{
+public class StudyMember{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    private String name;
     @CreatedDate
-    private LocalDateTime created_at;
-    private String description;
-    private Integer limit_members;
-    private LocalDateTime deadline;
-    private String status;
+    private LocalDateTime joined_at;
 
-    @ManyToOne
-    private User creator;
+    @ManyToOne(fetch = FetchType.LAZY)
+    private User user;
+    @ManyToOne(fetch = FetchType.LAZY)
+    private StudyGroup studyGroup;
 }
