+++ src/main/java/com/example/study_monster_back/tag/entity/StudyGroupTag.java
@@ -1,17 +1,23 @@
-package com.example.study_monster_back.entity;
+package com.example.study_monster_back.tag.entity;
+
+import com.example.study_monster_back.group.entity.StudyGroup;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
 import lombok.Data;
 
 @Entity
 @Data
-public class Tag{
+public class StudyGroupTag{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String name;
+    @ManyToOne
+    private Tag tag;
+    @ManyToOne
+    private StudyGroup studyGroup;
 }
