+++ src/main/java/com/example/study_monster_back/tag/entity/Tag.java
@@ -1,21 +1,17 @@
-package com.example.study_monster_back.entity;
+package com.example.study_monster_back.tag.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
 import lombok.Data;
 
 @Entity
 @Data
-public class StudyGroupTag{
+public class Tag{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne
-    private Tag tag;
-    @ManyToOne
-    private StudyGroup studyGroup;
+    private String name;
 }
